+++ arangodantic/graphs.py
@@ -65,6 +65,7 @@ class Graph(ABC):
 
         :raise UniqueConstraintViolated: Raised when there is a unique constraint
         violation.
+        :raise ModelNotFoundError: If any of the models are not found.
         """
 
         graph = cls.get_graph()
@@ -95,6 +96,8 @@ class Graph(ABC):
             except aioarangodb.exceptions.DocumentInsertError as ex:
                 if ex.error_code == ERROR_ARANGO_UNIQUE_CONSTRAINT_VIOLATED:
                     raise UniqueConstraintError(ex.error_message)
+                elif ex.error_code == ERROR_ARANGO_DOCUMENT_NOT_FOUND:
+                    raise ModelNotFoundError(ex.error_message)
                 raise
         else:
             # Update existing document
@@ -108,6 +111,8 @@ class Graph(ABC):
             except aioarangodb.exceptions.DocumentReplaceError as ex:
                 if ex.error_code == ERROR_ARANGO_UNIQUE_CONSTRAINT_VIOLATED:
                     raise UniqueConstraintError(ex.error_message)
+                elif ex.error_code == ERROR_ARANGO_DOCUMENT_NOT_FOUND:
+                    raise ModelNotFoundError(ex.error_message)
                 raise
 
         model.key_ = response["_key"]